Working on cleaning up code
Going through each method to clean them up a bit and trying to figure
out how to better link the tables and arrays to each other. Created a
method to overwrite array based on table values, still need to figure
out how to make the values editable without a lot of code.